import and loading of map packs
previous map pack is now in form of a zip file, instead of a folder

includes one map pack for usage with google, example how to use it can
be found in the MainActivity

importing is not completly finished und misses some error checks, but
can already be used.

currently only a complete map pack can be loaded.